+++ jwst/assign_wcs/nirspec.py
@@ -1554,9 +1554,10 @@ def gwa_to_ymsa(msa2gwa_model, lam_cen=None, slit=None, slit_y_range=None):
     return tab
 
 
-def nrs_wcs_set_input(input_model, slit_name, wavelength_range=None):
+def _nrs_wcs_set_input(input_model, slit_name):
     """
     Returns a WCS object for a specific slit, slice or shutter.
+    Does not compute the bounding box.
 
     Parameters
     ----------
@@ -1564,27 +1565,18 @@ def nrs_wcs_set_input(input_model, slit_name, wavelength_range=None):
         A WCS object for the all open slitlets in an observation.
     slit_name : int or str
         Slit.name of an open slit.
-    wavelength_range: list
-        Wavelength range for the combination of fliter and grating.
 
     Returns
     -------
     wcsobj : `~gwcs.wcs.WCS`
         WCS object for this slit.
     """
-    import copy  # TODO: Add a copy method to gwcs.WCS
+    import copy
     wcsobj = input_model.meta.wcs
-    if wavelength_range is None:
-        _, wrange = spectral_order_wrange_from_model(input_model)
-    else:
-        wrange = wavelength_range
+
     slit_wcs = copy.deepcopy(wcsobj)
     slit_wcs.set_transform('sca', 'gwa', wcsobj.pipeline[1].transform[1:])
-    # get the open slits from the model
-    # Need them to get the slit ymin,ymax
-    g2s = wcsobj.pipeline[2].transform
-    open_slits = g2s.slits
-
+    g2s = slit_wcs.pipeline[2].transform
     slit_wcs.set_transform('gwa', 'slit_frame', g2s.get_model(slit_name))
 
     exp_type = input_model.meta.exposure.type
@@ -1595,14 +1587,49 @@ def nrs_wcs_set_input(input_model, slit_name, wavelength_range=None):
     else:
         slit_wcs.set_transform('slit_frame', 'msa_frame',
                                wcsobj.pipeline[3].transform.get_model(slit_name) & Identity(1))
-    slit2detector = slit_wcs.get_transform('slit_frame', 'detector')
+    return slit_wcs
 
-    if is_nirspec_ifu:
-        bb = compute_bounding_box(slit2detector, wrange)
-    else:
+
+def nrs_wcs_set_input(input_model, slit_name, wavelength_range=None,
+                      slit_y_low=None, slit_y_high=None):
+    """
+    Returns a WCS object for a specific slit, slice or shutter.
+
+    Parameters
+    ----------
+    input_model : `~jwst.datamodels.DataModel`
+        A WCS object for the all open slitlets in an observation.
+    slit_name : int or str
+        Slit.name of an open slit.
+    wavelength_range: list
+        Wavelength range for the combination of filter and grating.
+
+    Returns
+    -------
+    wcsobj : `~gwcs.wcs.WCS`
+        WCS object for this slit.
+    """
+    def _get_y_range(input_model):
+        # get the open slits from the model
+        # Need them to get the slit ymin,ymax
+        g2s = input_model.meta.wcs.get_transform('gwa', 'slit_frame')
+        open_slits = g2s.slits
         slit = [s for s in open_slits if s.name == slit_name][0]
-        bb = compute_bounding_box(slit2detector, wrange,
-                                  slit_ymin=slit.ymin, slit_ymax=slit.ymax)
+        return slit.ymin, slit.ymax
+
+    if wavelength_range is None:
+        _, wavelength_range = spectral_order_wrange_from_model(input_model)
+
+    slit_wcs = _nrs_wcs_set_input(input_model, slit_name)
+    slit2detector = slit_wcs.get_transform('slit_frame', 'detector')
+    is_nirspec_ifu = is_nrs_ifu_lamp(input_model) or input_model.meta.exposure.type.lower() == 'nrs_ifu'
+    if is_nirspec_ifu:
+        bb = compute_bounding_box(slit2detector, wavelength_range)
+    else:
+        if slit_y_low is None or slit_y_high is None:
+            slit_y_low, slit_y_high = _get_y_range(input_model)
+        bb = compute_bounding_box(slit2detector, wavelength_range,
+                                  slit_ymin=slit_y_low, slit_ymax=slit_y_high)
 
     slit_wcs.bounding_box = bb
     return slit_wcs